escenario2/2_particiones/requirements.txt
grpcio
grpcio-tools
redis
